Tests the initial bot interface
