+++ bastion/tests/test_bastion.py
@@ -1,4 +1,6 @@
-from bastion import init_bastion
+from unittest.mock import patch, Mock
+
+from bastion import init_bastion, BastionBot
 from bastion.commands import Track, Misc, Roles
 
 def test_init_bastion():
@@ -11,6 +13,21 @@ def test_init_bastion():
 
     bastion = init_bastion(config)
 
-    assert bastion.http.token == token
-    for command in {Track, Misc, Roles}:
-        assert command in bastion.cogs.values()
+    assert bastion.token == token
+    assert {Track, Misc, Roles} == {cog.__class__ for cog in bastion.cogs.values()}
+
+
+def test_discord_run_is_called_with_token():
+    run_mock = Mock()
+    token = 'TOKEN'
+    config = {
+        'client': {
+            'token': token
+        }
+    }
+
+    bastion = init_bastion(config)
+    bastion.run = run_mock
+    bastion.run_bastion()
+
+    run_mock.assert_called_once_with(bastion.token)
